requirements.txt
# Azure
azure-ai-ml==1.12.1
azure-identity==1.15.0
azure-storage-blob==12.19.0
azure-mgmt-web==7.0.0
azure-mgmt-resource==23.0.1

# ML Framework
tensorflow-cpu==2.13.0
torch==2.0.1
pytorch-lightning==2.0.1
transformers==4.35.0
gluonts==0.13.4
prophet==1.1.4
xgboost==2.0.0

# ML Tools
mlflow==2.8.0
scikit-learn==1.3.0
pandas==2.1.0
numpy==1.24.3

# Web Framework
fastapi==0.95.2
uvicorn==0.24.0
python-multipart==0.0.6
typing-extensions==4.5.0

# Testing
pytest==7.4.0
pytest-asyncio==0.21.1
